requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
gunicorn==20.1.0
psycopg2-binary==2.9.9
